% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{categorisation}
\alias{categorisation}
\title{Title}
\usage{
categorisation(df, var_act, var_illus, var_grp)
}
\arguments{
\item{df}{a dataframe with}

\item{var_act}{un truc}

\item{var_illus}{un truc}

\item{var_grp}{un truc}
}
\description{
Title
}
