% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{charac_uni}
\alias{charac_uni}
\title{Univariate characterization of classes resulting from a clustering}
\usage{
charac_uni(obj)
}
\arguments{
\item{obj}{an object of type objcharac}
}
\value{
the cramers'V or the correlation for each variables and cluster depending on the variable's type
\describe{
\item{Case qualitative variables}{compute khi2, p-value, cramer's V}
\item{Case quantitative variables}{compute conditional means, correlation}
}
}
\description{
Univariate characterization of classes resulting from a clustering
}
\examples{
\dontrun{
data(insertion_master)
data.illu <- insertion_master[,c(1:6,12)]
data.act <-insertion_master[,7:11]
res.kmeans<-kmeans(data.act,centers=3,nstart=5)
obj <- objcharac(insertion_master, data.act, data.illu, res.kmeans$cluster)
uni = charac_uni(obj)
#show results for all variables
print(uni)
#example for the variable "domaine":
print(uni$domaine)
}


}
