% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviesbouldin.R
\name{db.categorisation}
\alias{db.categorisation}
\title{Compute Davies Bouldin's index}
\usage{
db.categorisation(obj)
}
\arguments{
\item{obj}{a categorisation}
}
\value{
the Davies Bouldin index
}
\description{
Compute Davies Bouldin's index
}
\examples{
X = c(10, 11, 12, 28)
Y = c(20, 23, 28, 15)
cl = c(1, 1, 1, 2)
data = as.data.frame(cbind(X, Y, cl))
colnames(data) = c("X", "Y", "cluster")
obj = categorisation(data, data[, -3], NULL, data$cluster)
db.categorisation(obj)

}
