% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daviesbouldin.R
\name{db_index}
\alias{db_index}
\title{Compute Davies Bouldin's index}
\usage{
db_index(obj)
}
\arguments{
\item{obj}{an object of type objcharac}
}
\value{
db_index returns the Davies Bouldin index
}
\description{
Function computing the Davies Bouldin's index which evaluate the clustering quality
}
\examples{
\dontrun{
X = c(10, 11, 12, 28)
Y = c(20, 23, 28, 15)
cl = c(1, 1, 1, 2)
data = as.data.frame(cbind(X, Y, cl))
colnames(data) = c("X", "Y", "cluster")
obj = categorisation(data, data[, -3], NULL, data$cluster)
db = db_index(obj)
print(db)
------------------
data(insertion_master)
data.illu <- insertion_master[,c(1:6,12)]
data.act <-insertion_master[,7:11]
res.kmeans<-kmeans(data.act,centers=3,nstart=5)
obj2 <- objcharac(insertion_master, data.act, data.illu, res.kmeans$cluster)
db2 = db_index(obj2)
print(db2)
}

}
