% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{objcharac}
\alias{objcharac}
\title{objcharac}
\usage{
objcharac(df, var_act, var_illus = NULL, var_grp)
}
\arguments{
\item{df}{a dataframe with all datas}

\item{var_act}{a dataframe with active variables}

\item{var_illus}{a dataframe with illustrative variables}

\item{var_grp}{a dataframe or a list with the clusters}
}
\value{
an object of type objcharac
\describe{
\item{$data}{dataframe}
\item{$act}{dataframe with active variables for clustering}
\item{$illus}{dataframe with illustrative variables}
\item{$grp}{clusters for each observations}
}
}
\description{
objcharac
}
\examples{
\dontrun{
library(charclust)
data(iris)
obj <- objcharac(iris, iris[,-5], NULL, iris$Species)
-----------------
data(insertion_master)
data.illu <- insertion_master[,c(1:6,12)]
data.act <-insertion_master[,7:11]
res.kmeans<-kmeans(data.act,centers=3,nstart=5)
obj2 <- objcharac(insertion_master, data.act, data.illu, res.kmeans$cluster)
}

}
