% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{sil.categorisation}
\alias{sil.categorisation}
\title{Compute silhouette information from clustering}
\usage{
sil.categorisation(obj)
}
\arguments{
\item{obj}{a categorisation}
}
\value{
a list containing the cluster means of silhouette, the total mean silhouette, a plot of cluster means of silhouette
}
\description{
Compute silhouette information from clustering
}
\examples{
X = c(10, 11, 12, 28)
Y = c(20, 23, 28, 15)
cl = c(1, 1, 1, 2)
data = as.data.frame(cbind(X, Y, cl))
colnames(data) = c("X", "Y", "cluster")
obj = categorisation(data, data[, -3], NULL, data$cluster)
s = sil.categorisation(obj)
s$silclus
s$silglob
s$plot

}
