% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{sil}
\alias{sil}
\title{Compute silhouette information from clustering}
\usage{
sil(obj)
}
\arguments{
\item{obj}{a categorisation}
}
\value{
sil returns a list containing:
\describe{
\item{$silclus}{the mean silhouette coefficients for each clusters}
\item{$silglob}{the total mean silhouette coefficient}
\item{$plot}{a plot of cluster means of silhouette}
}
}
\description{
Function computing the silhouette which evaluate the clustering quality
}
\examples{
\dontrun{
X = c(10, 11, 12, 28)
Y = c(20, 23, 28, 15)
cl = c(1, 1, 1, 2)
data = as.data.frame(cbind(X, Y, cl))
colnames(data) = c("X", "Y", "cluster")
obj = categorisation(data, data[, -3], NULL, data$cluster)
s = sil(obj)
s$silclus
s$silglob
s$plot
------------------
data(insertion_master)
data.illu <- insertion_master[,c(1:6,12)]
data.act <-insertion_master[,7:11]
res.kmeans<-kmeans(data.act,centers=3,nstart=5)
obj2 <- objcharac(insertion_master, data.act, data.illu, res.kmeans$cluster)
s2 = sil(obj2)
s2$silclus
s2$silglob
s2$plot
}

}
