% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariee.categorisation}
\alias{univariee.categorisation}
\title{Univariate characterization of classes resulting from a clustering}
\usage{
univariee.categorisation(obj)
}
\arguments{
\item{obj}{a categorisation}
}
\value{
the cramers'V or the correlation for each variables and cluster depending on the variable's type
}
\description{
Univariate characterization of classes resulting from a clustering
}
