% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{categorisation}
\alias{categorisation}
\title{Categorisation}
\usage{
categorisation(df, var_act, var_illus, var_grp)
}
\arguments{
\item{df}{a dataframe with all datas}

\item{var_act}{a dataframe with active variables}

\item{var_illus}{a dataframe with illustrative variables}

\item{var_grp}{a dataframe with the clusters}
}
\value{
a categorisation object
}
\description{
Categorisation
}
\examples{
library(charclust)
data(iris)
obj = categorisation(iris, iris[,-5], NULL, iris$Species)

}
